+++ v3/pkg/application/options_application.go
@@ -51,6 +51,11 @@ type Options struct {
 
 	// KeyBindings is a map of key bindings to functions
 	KeyBindings map[string]func(window *WebviewWindow)
+
+	// OnShutdown is called when the application is about to quit.
+	// This is useful for cleanup tasks.
+	// The shutdown process blocks until this function returns
+	OnShutdown func()
 }
 
 // AssetOptions defines the configuration of the AssetServer.